feat: Complete website layout
